pyproject.toml
[project]
name = "unsplash-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=0.4.1",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.1",
]
